test(segment): migrate icon
